[feature] add support for better scopes
Create a common utils library for -extra specific code, and update
perfwrapper to use it.
