--- src/perfwrapper/Makefile.am
@@ -1,6 +1,0 @@
-AM_CFLAGS = @LIBNRM_CFLAGS@ @PAPI_CFLAGS@
-AM_LDFLAGS = @LIBNRM_LIBS@ @PAPI_LIBS@
-
-nrm_perfwrapper_SOURCES = perfwrapper.c
-
-bin_PROGRAMS = nrm-perfwrapper
